Minor auto-refactor code cleanup on a massive scale (#354)
* Minor auto-refactor code cleanup on a massive scale

* Remove unnecessary cast

+++ OpenLdapSync/src/org/labkey/openldapsync/ldap/LdapEntry.java
@@ -75,8 +75,8 @@ public class LdapEntry
 
         try
         {
-            Integer value = Integer.parseInt(a);
-            return (value.intValue() & 2) == 0;
+            int value = Integer.parseInt(a);
+            return (value & 2) == 0;
         }
         catch (NumberFormatException e)
         {
